jarvis.py/requirements.txt
speechrecognition
pyttsx
pyaudio